Move useGetCurrentOrganizationType under admin.core.v1/hooks
