+++ features/admin.applications.v1/components/forms/inbound-oidc-form.tsx
@@ -57,7 +57,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { Dispatch } from "redux";
 import { Button, Container, Divider, DropdownProps, Form, Grid, Label, List, Table } from "semantic-ui-react";
 import { AppState, ConfigReducerStateInterface } from "../../../admin.core.v1";
-import { useGlobalVariablesContext } from "../../../admin.core.v1/context/global-variables-context";
+import { useGlobalVariablesContext } from "../../../admin.core.v1/hooks/use-global-variables";
 import { applicationConfig } from "../../../admin.extensions.v1";
 import { getSharedOrganizations } from "../../../admin.organizations.v1/api";
 import { OrganizationType } from "../../../admin.organizations.v1/constants";
