+++ features/admin.applications.v1/components/modals/application-share-modal.tsx
@@ -55,7 +55,7 @@ import {
     Transition
 } from "semantic-ui-react";
 import { AppState, EventPublisher } from "../../../admin.core.v1";
-import { useGlobalVariablesContext } from "../../../admin.core.v1/context/global-variables-context";
+import { useGlobalVariablesContext } from "../../../admin.core.v1/hooks/use-global-variables";
 import {
     getOrganizations,
     getSharedOrganizations,
