+++ features/admin.core.v1/hooks/use-routes.tsx
@@ -23,7 +23,7 @@ import isEmpty from "lodash-es/isEmpty";
 import { useDispatch, useSelector } from "react-redux";
 import { Dispatch } from "redux";
 import useAuthorization from "../../admin.authorization.v1/hooks/use-authorization";
-import { useGlobalVariablesContext } from "../../admin.core.v1/context/global-variables-context";
+import { useGlobalVariablesContext } from "../../admin.core.v1/hooks/use-global-variables";
 import { useGetCurrentOrganizationType } from "../../admin.organizations.v1/hooks/use-get-organization-type";
 import { getAppViewRoutes } from "../configs/routes";
 import { AppConstants } from "../constants/app-constants";
